+++ .github/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 (Dates are in YYYY-MM-DD format. This message is mainly for my own sake.)
 
 ## [Unreleased]
+### Added
+* Action log
+  * Alongside `actionLogChannel` config parameter
 
 ## [4.2.0] - 2021-05-05
 ### Added